Just added a new file

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -5,7 +5,7 @@
  * Return:void
  */
 
-void print_number(int n)
+void print_number(long int n)
 {
 
 	int d, i, ii;
